moved default materials out of the aframe portion of the code
