+++ src/aframe/component/architecture-toolkit/column.js
@@ -7,6 +7,7 @@ import generateNormals from '../../../utils/data3d/buffer/get-normals'
 import generateUvs from '../../../utils/data3d/buffer/get-uvs'
 import cloneDeep from 'lodash/cloneDeep'
 import columnData3d from '../../../scene/structure/parametric-objects/column'
+import dataToMaterials from './common/data-to-materials'
 
 export default {
 
@@ -26,31 +27,8 @@ export default {
     // get defaults and
     var attributes = cloneDeep(data)
 
-
-    // setup materials
-    // defaults
-    var materials = {
-      top: 'wall_top',
-      side: 'basic-wall'
-    }
-
-    // check for adapted materials
-    var materialKeys = Object.keys(data).filter(function(key) {
-      return key.indexOf('material_') > -1
-    })
-    // add materials to instance
-    materialKeys.forEach(function(key) {
-      var mesh = key.replace('material_', '')
-      materials[mesh] = data[key]
-    })
-
-    // fetch materials from mat library
-    Object.keys(materials).forEach(mat => {
-      materials[mat] = getMaterial(materials[mat])
-    })
-
-    attributes.materials = materials
-
+    attributes.materials = dataToMaterials(data)
+    
     // construct data3d object
 
     var data3d = columnData3d(attributes)
