+++ src/aframe/component/architecture-toolkit/door.js
@@ -6,6 +6,7 @@ import updateSchema from './common/update-schema.js'
 import generateNormals from '../../../utils/data3d/buffer/get-normals'
 import cloneDeep from 'lodash/cloneDeep'
 import doorData3d from '../../../scene/structure/parametric-objects/door'
+import dataToMaterials from './common/data-to-materials'
 
 export default {
 
@@ -49,30 +50,7 @@ export default {
 
     let attributes = cloneDeep(data)
 
-    // setup materials
-    // defaults
-    var materials = {
-      frame: {
-        colorDiffuse: [0.95, 0.95, 0.95],
-        colorSpecular: [0.04, 0.04, 0.04],
-        specularCoef: 30
-      },
-      leaf: 'doorLeaf-flush-white',
-      handle: 'aluminium',
-      threshold: 'wood_parquet_oak'
-    }
-
-    // check for adapted materials
-    var materialKeys = Object.keys(data).filter(function(key) {
-      return key.indexOf('material_') > -1
-    })
-    // add materials to instance
-    materialKeys.forEach(function(key) {
-      var mesh = key.replace('material_', '')
-      materials[mesh] = data[key]
-    })
-
-    attributes.materials=materials
+    attributes.materials = dataToMaterials(data)
 
     // construct data3d object
     var data3d = await doorData3d(attributes)
