+++ src/aframe/component/architecture-toolkit/kitchen.js
@@ -5,6 +5,7 @@ import getSchema from './common/get-schema.js'
 import updateSchema from './common/update-schema.js'
 import cloneDeep from 'lodash/cloneDeep'
 import kitchenData3d from '../../../scene/structure/parametric-objects/kitchen'
+import dataToMaterials from './common/data-to-materials'
 
 export default {
 
@@ -24,33 +25,7 @@ export default {
     // get defaults and
     let attributes = cloneDeep(data)
 
-    // setup materials
-    // defaults
-    let materials = {
-      kitchen: 'cabinet_paint_white',
-      counter: 'counter_granite_black',
-      tab: 'chrome',
-      oven: 'oven_miele_60-60',
-      cooktop: 'cooktop_westinghouse_60',
-      microwave: 'microwave_samsung',
-      chrome: 'chrome',
-      black_metal: {
-        "specularCoef": 24,
-        "colorDiffuse": [0.02, 0.02, 0.02],
-        "colorSpecular": [0.7, 0.7, 0.7]
-      }
-    }
-
-    // check for adapted materials
-    var materialKeys = Object.keys(data).filter(function(key) {
-      return key.indexOf('material_') > -1
-    }).forEach(function(key) {
-      // add materials to instance
-      var mesh = key.replace('material_', '')
-      materials[mesh] = data[key]
-    })
-
-    attributes.materials=materials;
+    attributes.materials = dataToMaterials(data)
 
     // get meshes and materials
     // promised base because it loads external meshes
