+++ src/aframe/component/architecture-toolkit/railing.js
@@ -6,6 +6,7 @@ import updateSchema from './common/update-schema.js'
 import generateNormals from '../../../utils/data3d/buffer/get-normals'
 import cloneDeep from 'lodash/cloneDeep'
 import wallRailing3d from '../../../scene/structure/parametric-objects/railing'
+import dataToMaterials from './common/data-to-materials'
 
 export default {
 
@@ -25,30 +26,7 @@ export default {
     // get defaults and
     let attributes = cloneDeep(data)
 
-    // setup materials
-    // defaults
-    var materials = {
-      railing: {
-        colorDiffuse: [0.85, 0.85, 0.0]
-      }
-    }
-
-    // check for adapted materials
-    var materialKeys = Object.keys(data).filter(function(key) {
-      return key.indexOf('material_') > -1
-    })
-    // add materials to instance
-    materialKeys.forEach(function(key) {
-      var mesh = key.replace('material_', '')
-      materials[mesh] = data[key]
-    })
-
-    // fetch materials from mat library
-    Object.keys(materials).forEach(mat => {
-      materials[mat] = getMaterial(materials[mat])
-    })
-
-    attributes.materials = materials;
+    attributes.materials = dataToMaterials(data)
 
     // construct data3d object
     let data3d = await wallRailing3d(attributes)
