+++ src/scene/structure/parametric-objects/closet.js
@@ -10,12 +10,18 @@ import getMaterials3d from './common/get-materials'
 export default function getData3d(attributes) {
     return Promise.all([
       generateMeshes3d(attributes),
-      getMaterials3d(attributes.materials)
+      getMaterials3d(attributes.materials, getDefaultMaterials())
     ]).then(results => ({
       meshes: results[0],
       materials: results[1]
     }))
   }
+
+export function getDefaultMaterials(){
+  return {
+    closet: 'cabinet_paint_white'
+  }
+}
 
 export function generateMeshes3d(a) {
