+++ src/scene/structure/parametric-objects/polyfloor.js
@@ -11,11 +11,19 @@ import getMaterials3d from './common/get-materials'
 export default function getData3d(attributes) {
   return Promise.all([
     generateMeshes3d(attributes),
-    getMaterials3d(attributes.materials)
+    getMaterials3d(attributes.materials, getDefaultMaterials())
   ]).then(results => ({
     meshes: results[0],
     materials: results[1]
   }))
+}
+
+export function getDefaultMaterials(){
+  return  {
+    top: 'wood_parquet_oak',
+    side: 'basic-wall',
+    ceiling: 'basic-ceiling'
+  }
 }
 
 export function generateMeshes3d (a) {
