+++ src/scene/structure/parametric-objects/railing.js
@@ -13,11 +13,19 @@ import loadData3d from '../../../utils/data3d/load'
 export default async function getData3d(attributes) {
   return Promise.all([
     generateMeshes3d(attributes),
-    getMaterials3d(attributes.materials)
+    getMaterials3d(attributes.materials, getDefaultMaterials())
   ]).then(results => ({
     meshes: results[0],
     materials: results[1]
   }))
+}
+
+export function getDefaultMaterials(){
+  return {
+    railing: {
+      colorDiffuse: [0.85, 0.85, 0.0]
+    }
+  }
 }
 
 export function generateMeshes3d (a) {
