+++ src/scene/structure/parametric-objects/stairs.js
@@ -9,11 +9,21 @@ import getMaterials3d from './common/get-materials'
 export default function getData3d(attributes) {
   return Promise.all([
     generateMeshes3d(attributes),
-    getMaterials3d(attributes.materials)
+    getMaterials3d(attributes.materials, getDefaultMaterials())
   ]).then(results => ({
     meshes: results[0],
     materials: results[1]
   }))
+}
+
+export function getDefaultMaterials(){
+  return {
+    steps: 'basic-wall',
+    tread: 'wood_parquet_oak',
+    railing: {
+      colorDiffuse: [0.85, 0.85, 0.85]
+    }
+  }
 }
 
 export function generateMeshes3d (a) {
